Update arcade-playerrecords-design.html
Included various entries in order to simulate a larger leaderboard.

Added a title with the <h2> tag.

Separated the span classes with the button with <p> so that they are displayed separately. With this the button is fixed on the second line .

Changed the input tag to the button tag and added an id in order to design the button easier.